requirements.txt
asgiref==3.2.10
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
coverage==4.0.3
Django==3.0.7
django-cors-headers==3.4.0
django-crispy-forms==1.9.1
django-extensions==3.0.2
django-filters==0.2.1
djangorestframework==3.11.0
drf-yasg==1.17.1
idna==2.9
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.4.0
packaging==20.4
Pillow==7.1.2
pluggy==0.13.1
psycopg2==2.8.5
py==1.8.2
pyparsing==2.4.7
pytest==5.4.3
pytest-cov==2.10.0
pytest-django==3.9.0
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
soft-delete==0.2.2
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.2.4
